refactor(library)!: remove deprecated apis
