+++ javascript-modules-library/src/components/AbsoluteArea.tsx
@@ -11,13 +11,10 @@ import type { JCRNodeWrapper } from "org.jahia.services.content";
 export function AbsoluteArea({
   name,
   parent,
-  areaView,
   view,
-  allowedTypes,
   allowedNodeTypes,
   numberOfItems,
   readOnly = false,
-  areaType = "jnt:contentList",
   nodeType = "jnt:contentList",
   parameters,
 }: Readonly<{
@@ -29,20 +26,8 @@ export function AbsoluteArea({
    */
   parent: JCRNodeWrapper;
 
-  /**
-   * The view to use for the area.
-   *
-   * @deprecated Use {@link #view} instead
-   */
-  areaView?: string;
   /** The view to use for the area. */
   view?: string;
-  /**
-   * The allowed types for the area.
-   *
-   * @deprecated Use {@link #allowedNodeTypes} instead
-   */
-  allowedTypes?: string[];
   /** The allowed types for the area. */
   allowedNodeTypes?: string[];
   /** The number of items to display in the area. */
@@ -62,21 +47,10 @@ export function AbsoluteArea({
   /**
    * Node type to be used to create the area (if the node does not exist yet)
    *
-   * @deprecated Use {@link #nodeType} instead
-   * @default jnt:contentList
-   */
-  areaType?: string;
-  /**
-   * Node type to be used to create the area (if the node does not exist yet)
-   *
    * @default jnt:contentList
    */
   nodeType?: string;
-  /**
-   * Map of custom parameters that can be passed to the backend engine for advanced logic.
-   *
-   * @deprecated Not recommended
-   */
+  /** Map of custom parameters that can be passed to the backend engine for advanced logic. */
   parameters?: Record<string, unknown>;
 }>): React.JSX.Element {
   const { renderContext } = useServerContext();
@@ -88,10 +62,10 @@ export function AbsoluteArea({
           {
             name,
             parent: parent,
-            view: view ? view : areaView,
-            allowedNodeTypes: allowedNodeTypes ?? allowedTypes,
+            view,
+            allowedNodeTypes,
             numberOfItems,
-            nodeType: nodeType ?? areaType,
+            nodeType,
             editable: readOnly !== true,
             limitedAbsoluteAreaEdit: readOnly === "children",
             parameters,
